LCOW: add "no_lcow" build tag to allow disabling lcow
